+++ egui/src/containers/mod.rs
@@ -4,6 +4,7 @@
 
 pub(crate) mod area;
 pub(crate) mod collapsing_header;
+mod combo_box;
 pub(crate) mod frame;
 pub(crate) mod popup;
 pub(crate) mod resize;
@@ -11,6 +12,6 @@ pub(crate) mod scroll_area;
 pub(crate) mod window;
 
 pub use {
-    area::Area, collapsing_header::CollapsingHeader, frame::Frame, popup::*, resize::Resize,
-    scroll_area::ScrollArea, window::Window,
+    area::Area, collapsing_header::CollapsingHeader, combo_box::*, frame::Frame, popup::*,
+    resize::Resize, scroll_area::ScrollArea, window::Window,
 };