+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Color picker
 * Unicode characters in labels (limited by [what the default font supports](https://fonts.google.com/specimen/Comfortaa#glyphs))
+* Simple drop-down combo box menu
 
 ## 0.1.4 - 2020-09-08